project restructuring
move views inside /views/ package
